webserver: reorganize and refactor javascript
Creates a series of classes, 1 per page, and a base class. Moves
common functions to various modules.
